modify streamWebRTC
